add Explain Activity and collect common functions to BaseActivity
