+++ app/src/main/java/me/rei_m/hbfavmaterial/activities/BaseActivity.kt
@@ -3,25 +3,19 @@ package me.rei_m.hbfavmaterial.activities
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import android.support.v7.widget.Toolbar
-import com.squareup.otto.Subscribe
+import android.view.MenuItem
 import me.rei_m.hbfavmaterial.R
 import me.rei_m.hbfavmaterial.events.EventBusHolder
-import me.rei_m.hbfavmaterial.events.UserIdCheckedEvent
-import me.rei_m.hbfavmaterial.extensions.setFragment
-import me.rei_m.hbfavmaterial.extensions.startActivityWithClearTop
-import me.rei_m.hbfavmaterial.fragments.SplashFragment
 
-public class SplashActivity : AppCompatActivity() {
+public abstract class BaseActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_splash)
+        setContentView(R.layout.activity)
         val toolbar = findViewById(R.id.toolbar) as Toolbar
         setSupportActionBar(toolbar)
-
-        if (savedInstanceState == null) {
-            setFragment(SplashFragment.newInstance())
-        }
+        supportActionBar.setDisplayHomeAsUpEnabled(true)
+        supportActionBar.setHomeButtonEnabled(true)
     }
 
     override fun onResume() {
@@ -38,11 +32,17 @@ public class SplashActivity : AppCompatActivity() {
         EventBusHolder.EVENT_BUS.unregister(this)
     }
 
-    @Subscribe
-    @SuppressWarnings("unused")
-    public fun onUserIdChecked(event: UserIdCheckedEvent) {
-        if (event.type == UserIdCheckedEvent.Companion.Type.OK) {
-            startActivityWithClearTop(MainActivity.createIntent(applicationContext))
+    override fun onOptionsItemSelected(item: MenuItem?): Boolean {
+
+        val id = item?.itemId
+
+        when (id) {
+            android.R.id.home ->
+                finish()
+            else ->
+                return super.onOptionsItemSelected(item)
         }
+
+        return true
     }
 }
