+++ app/src/main/java/me/rei_m/hbfavmaterial/activities/BookmarkActivity.kt
@@ -5,8 +5,6 @@ import android.content.Intent
 import android.os.Bundle
 import android.support.design.widget.FloatingActionButton
 import android.support.v4.app.ShareCompat
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.Toolbar
 import android.view.Menu
 import android.view.MenuItem
 import com.squareup.otto.Subscribe
@@ -16,13 +14,12 @@ import me.rei_m.hbfavmaterial.entities.EntryEntity
 import me.rei_m.hbfavmaterial.events.BookmarkClickedEvent
 import me.rei_m.hbfavmaterial.events.BookmarkCountClickedEvent
 import me.rei_m.hbfavmaterial.events.BookmarkUserClickedEvent
-import me.rei_m.hbfavmaterial.events.EventBusHolder
 import me.rei_m.hbfavmaterial.extensions.replaceFragment
 import me.rei_m.hbfavmaterial.extensions.setFragment
 import me.rei_m.hbfavmaterial.fragments.BookmarkFragment
 import me.rei_m.hbfavmaterial.fragments.EntryWebViewFragment
 
-public class BookmarkActivity : AppCompatActivity() {
+public class BookmarkActivity : BaseActivity() {
 
     private var mEntryTitle: String = ""
     private var mEntryLink: String = ""
@@ -50,11 +47,6 @@ public class BookmarkActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity)
-        val toolbar = findViewById(R.id.toolbar) as Toolbar
-        setSupportActionBar(toolbar)
-        supportActionBar.setDisplayHomeAsUpEnabled(true)
-        supportActionBar.setHomeButtonEnabled(true)
 
         if (savedInstanceState == null) {
             if (intent.hasExtra(ARG_BOOKMARK)) {
@@ -81,20 +73,6 @@ public class BookmarkActivity : AppCompatActivity() {
 
     }
 
-    override fun onResume() {
-        super.onResume()
-
-        // EventBus登録
-        EventBusHolder.EVENT_BUS.register(this)
-    }
-
-    override fun onPause() {
-        super.onPause()
-
-        // EventBus登録解除
-        EventBusHolder.EVENT_BUS.unregister(this)
-    }
-
     override fun onRestoreInstanceState(savedInstanceState: Bundle?) {
         super.onRestoreInstanceState(savedInstanceState)
         mEntryTitle = savedInstanceState?.getString(KEY_ENTRY_TITLE)!!
@@ -118,8 +96,6 @@ public class BookmarkActivity : AppCompatActivity() {
         val id = item?.itemId;
 
         when (id) {
-            android.R.id.home ->
-                finish();
             R.id.menu_share ->
                 ShareCompat.IntentBuilder.from(this)
                         .setChooserTitle("記事をシェアします")
